Alerting edits
I included one alert that is attached to the section "Applying an affix." This alert informs users that a horse can only have one affix, and a player can only assign an affix to a horse if that horse was born in that player's breeding farm. Otherwise, a player may not remove or change any affixes that have already been assigned to a horse by that horse's original breeder. This is a significant piece of information for a player to know.