Subtitle edits: Removed gerund
I don't know if this is allowed, and I'll go back and double check our readings, but I kind of like the use of "How to ____" as the beginning phrase for the subtitles. I'll keep the use of the gerund in the main title (I did see in a reading that we should do that). But the subtitles beginning with "How to ____" makes more sense to me. It also distinguishes them as the actual "coaching" section(s) apart from the About and Prereqs sections.