Staging edits (pt. 2)
I cleaned up some of the staging elements of this procedure to, hopefully, only include information that was valuable to the player and that does not accidentally leave out important things players should consider within the context of this procedure. There are not really any prerequisites for this procedure, but I think what I put in that section here is pretty important for players to know.