Finalized adding in images, and made suggested changes
I added in all the necessary screenshots.

I also reworked the color selection instructions to make a bit more sense.

I reworded my explanation of the favorite thing to be clearer that it has no impact on game play.

I emphasized the section on which aspects can be changed in game.

Improved the pre-reqs to be clearer as to what steam is/why you need to be logged in