Added images and tweaks from peer review
I added in images to all of my steps.
I also changed the pre-requisites to match the wording from the other instructions to ensure that the role of Steam was clear.

I also added in an image, and reworked the wording about clearing the plot so that it was clearer to the reader.

I moved the sprinkler/fertilizer section to the end as it is option and I don't want to distract the user too much from the task.