Resetting the Room, Console Connection outlining
Adding a bulleted list to the room resetting to serve as a checklist for users instead of another ordered list of tasks to complete, which can seem mentally taxing after setting up and running an event all day. I also added information on lost and found policies in the event of a fail state, in this case would be misplacing personal belongings in the Landing.